# openapi.yaml
openapi: 3.1.0
info:
  title: Docker Visualizer API
  description: |
    Read-only Docker inventory and visualization API.
    Metrics units are explicit (see schemas). Volume sizes use availability semantics (ADR-011).
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Default local bind (ADR-009)
...
paths:
...
  /hosts:
    get:
      tags: [System]
      summary: List configured Docker hosts (ADR-014)
      operationId: listHosts
      responses:
        "200":
          description: Host registry
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, defaultHost, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  defaultHost: { type: string }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HostInfo"
        "503":
          description: Hosts not initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /containers:
    get:
      tags: [Containers]
      summary: List containers from inventory snapshot
      parameters:
        - $ref: "#/components/parameters/HostQuery"
        - name: stack
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: health
          in: query
          schema: { type: string }
        - name: q
          in: query
          description: Search name/image/id/stack/service
          schema: { type: string }
      responses:
        "200":
          description: Container list
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  snapshotAt:
                    type: [string, "null"]
                    format: date-time
                  snapshotAgeMs: { type: integer, format: int64 }
                  collectError:
                    type: [string, "null"]
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Container"
  /containers/{id}:
    get:
      tags: [Containers]
      summary: Get container by full or short ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Container details
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  snapshotAt:
                    type: [string, "null"]
                    format: date-time
                  data:
                    $ref: "#/components/schemas/Container"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /containers/{id}/stats:
    get:
      tags: [Containers]
      summary: Latest cached stats sample (Docker CLI-compatible formulas)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Stats sample; data may be null if stopped or not yet sampled
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, running, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  snapshotAt:
                    type: [string, "null"]
                    format: date-time
                  statsAt:
                    type: [string, "null"]
                    format: date-time
                  running: { type: boolean }
                  data:
                    oneOf:
                      - $ref: "#/components/schemas/ContainerStats"
                      - type: "null"
        "404":
          description: Container not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /containers/{id}/inspect:
    get:
      tags: [Containers]
      summary: Live Docker inspect JSON (redacted by default)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: redact
          in: query
          description: Default true — strips Env, RegistryAuth, secret-looking labels
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Inspect document
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  data:
                    type: object
                    required: [id, redacted, inspect]
                    properties:
                      id: { type: string }
                      name: { type: string }
                      redacted: { type: boolean }
                      redactedFields:
                        type: array
                        items: { type: string }
                      inspect:
                        type: object
                        additionalProperties: true
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /containers/{id}/logs:
    get:
      tags: [Containers]
      summary: On-demand log snapshot (not persisted)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: tail
          in: query
          schema: { type: integer, default: 200 }
        - name: since
          in: query
          schema: { type: string }
        - name: timestamps
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Demultiplexed log text
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  data:
                    type: object
                    required: [id, text, truncated, warning]
                    properties:
                      id: { type: string }
                      name: { type: string }
                      tail: { type: integer }
                      since: { type: [string, "null"] }
                      timestamps: { type: boolean }
                      truncated: { type: boolean }
                      text: { type: string }
                      warning: { type: string }
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /containers/{id}/logs/ws:
    get:
      tags: [Containers]
      summary: Follow log stream (WebSocket)
      operationId: streamContainerLogs
      description: |
        Upgrade to WebSocket. Server frames:
        `container.logs.status`, `container.logs` ({ text }), `error`.
        Auth: Bearer header or `?access_token=` (same as hub).
        Logs are not persisted.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: tail
          in: query
          schema: { type: integer, default: 200 }
        - name: timestamps
          in: query
          schema: { type: boolean, default: false }
        - name: since
          in: query
          schema: { type: string }
        - name: access_token
          in: query
          required: false
          schema: { type: string }
      responses:
        "101":
          description: Switching Protocols
        "401":
          description: Missing or invalid auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Container not found (sent as WS error frame after upgrade when applicable)
...
  /stacks/{name}:
    get:
      tags: [Stacks]
      summary: Get one stack
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Stack detail
        "404":
          description: Not found
  /stacks/{name}/volumes:
    get:
      tags: [Stacks]
      summary: Volumes used by a stack
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Stack volumes
        "404":
          description: Not found
  /export:
    get:
      tags: [System]
      summary: Download inventory export (JSON/CSV)
      operationId: exportInventory
      description: |
        Structured PowerShell-replacement export using the parity schema (schemaVersion 1).
        JSON includes containers, stacks, and totals. CSV is flat tables.
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [json, csv], default: json }
        - name: scope
          in: query
          description: |
            all | containers | stacks.
            For CSV, `all` exports the containers table.
          schema: { type: string, enum: [all, containers, stacks], default: all }
      responses:
        "200":
          description: File download
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
            text/csv:
              schema: { type: string }
          headers:
            Content-Disposition:
              schema: { type: string }
        "400":
          description: Bad format/scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /graph:
    get:
      tags: [System]
      summary: Topology graph (stack → service → container → net/vol/image)
      parameters:
        - name: scope
          in: query
          description: all | stack
          schema: { type: string, default: all }
        - name: stack
          in: query
          description: Required when scope=stack
          schema: { type: string }
      responses:
        "200":
          description: Graph nodes and edges
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  snapshotAt:
                    type: [string, "null"]
                    format: date-time
                  statsAt:
                    type: [string, "null"]
                    format: date-time
                  snapshotAgeMs: { type: integer, format: int64 }
                  data:
                    $ref: "#/components/schemas/Graph"
        "400":
          description: Invalid scope/stack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /networks:
    get:
      tags: [Networks]
      summary: List networks
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: driver
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Network list
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  snapshotAt:
                    type: [string, "null"]
                    format: date-time
                  snapshotAgeMs: { type: integer, format: int64 }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Network"
  /networks/{id}:
    get:
      tags: [Networks]
      summary: Get network by id, short id, or name
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Network detail
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  data:
                    $ref: "#/components/schemas/Network"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /volumes:
    get:
      tags: [Volumes]
      summary: List volumes with usage availability
      parameters:
        - name: stack
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Volume list
  /volumes/{name}:
    get:
      tags: [Volumes]
      summary: Get one volume
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Volume detail
        "404":
          description: Not found
  /images:
    get:
      tags: [Images]
      summary: List images
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: dangling
          in: query
          description: '"true" | "false" filter'
          schema: { type: string }
      responses:
        "200":
          description: Image list
          content:
            application/json:
              schema:
                type: object
                required: [timestamp, data]
                properties:
                  timestamp: { type: string, format: date-time }
                  snapshotAt:
                    type: [string, "null"]
                    format: date-time
                  snapshotAgeMs: { type: integer, format: int64 }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Image"
...
  /provenance:
    get:
      tags: [Provenance]
      summary: List data provenance specs
      operationId: listProvenance
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Provenance catalog
...
components:
  parameters:
    HostQuery:
      name: host
      in: query
      required: false
      schema: { type: string }
      description: Named Docker host (ADR-014). Omit for the process default.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Required when `--auth-token` is set (always required for non-loopback listen).
        WebSocket clients may use `?access_token=` instead.
  schemas:
    HostInfo:
      type: object
      required: [name, endpoint, source, isDefault, connected]
      properties:
        name: { type: string }
        endpoint: { type: string }
        source: { type: string }
        context: { type: string }
        isDefault: { type: boolean }
        connected: { type: boolean }
        error: { type: string }
...
    ErrorResponse:
      type: object
      required: [error, timestamp]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
...